Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, FlatList,Text } from 'react-native';
+import { View, FlatList, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -34,7 +34,24 @@ const mapStateToProps = state => {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress:() => this.props.deleteFavorite(item.id)
+                    onPress:() => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            `Are You Sure You Wish To Delete The Favorite Dish${item.name}?`,
+                            [
+                                {
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'ok',
+                                    onPress:() => this.props.deleteFavorite(item.id) 
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
